setup/requirements.txt
thgcmd
getmac
colorama==0.4.1
attrs==19.1.0
wcwidth==0.1.7
pyperclip==1.7.0
art==3.0
tabulate==0.8.3
requests==2.21.0
python-pacman
sqlalchemy
docker